rest: allow reading byte range from a block
This feature will allow to building external indexes, which support
directly fetching a specific transaction (similar to TxIndex::FindTx).

Today, electrs and other indexers are used to maintain a map between
an address/scripthash [1] to the list of relevant transactions.
However, when fetching those transactions from bitcoind, they rely on
reading the whole block (and post-filtering for specific transaction)
or using txindex to fetch a transaction using its txid.

This PR is adding support for using the transaction's byte offset
within its block to be able to fetch it directly using REST API.

If binary response format is used, the transaction data will be read
directly from the storage and sent back to the client, without any
deserialization overhead.

IIUC, this resolves the concerns raised in [2], since it doesn't
rely on exposing internal bitcoind file format to an external indexer,
while providing an efficient way to retrieve any specific transaction.

[1] https://github.com/romanz/electrs/blob/master/doc/schema.md
[2] https://github.com/bitcoin/bitcoin/pull/20702