index: store per-block transaction locations for efficient lookups
Today, electrs and other indexers are used to maintain a map between
an address/scripthash to the list of the relevant transactions.

However, when fetching those transactions from bitcoind, they rely on
reading the whole block (and post-filtering for specific transaction)
[1] or using a `txindex` to fetch a transaction via its txid [2,3,4].

The above approach has significant storage and CPU overhead, since
the `txid` is a pseudo-random 32-byte value.

This PR is adding support for using the transaction's offset within
its block to be able to read it directly using REST API.

If binary response format is used, the transaction data will be read
directly from the storage and sent back to the client, without any
deserialization overhead.

The resulting index is much smaller (allowing it to be cached):

  $ du -sh indexes/offsets/ indexes/txindex/
  2.3G  indexes/offsets/
  51G   indexes/txindex/

[1] https://github.com/romanz/electrs/blob/master/doc/schema.md
[2] https://github.com/Blockstream/electrs/blob/new-index/doc/schema.md#txstore
[3] https://github.com/spesmilo/electrumx/blob/master/docs/HOWTO.rst#prerequisites
[4] https://github.com/cculianu/Fulcrum/blob/master/README.md#requirements