index: store per-block transactions' byte offsets
This feature will allow to building external indexes, which require
directly fetching specific transactions.

Today, electrs and other indexers are used to maintain a map between
an address/scripthash to the list of relevant blocks/transactions.
However, when fetching those transactions from bitcoind, they rely on
reading the whole block (and post-filtering for specific transaction)
[1] or using `txindex` to fetch a transaction via its txid [2, 3].

This PR is adding support for using the transaction's position
within its block to be able to read it directly using REST API.

It would be much more efficient to encode and store the position in
an external index data structure - most of the blocks have at most
a few thousand transactions, so on average we will need ~2 bytes to
represent transaction position within the block's transactions.

If binary response format is used, the transaction data will be read
directly from the storage and sent back to the client, without any
deserialization overhead.

IIUC, this approach addresses the concerns raised in [4], since it
doesn't expose internal `bitcoind` file format to an external indexer,
while providing an efficient way to retrieve any specific transaction.

The additional index supports pruning, since it stores the transactions'
byte offsets within each block.

[1] https://github.com/romanz/electrs/blob/master/doc/schema.md
[2] https://github.com/spesmilo/electrumx/blob/master/docs/HOWTO.rst#prerequisites
[3] https://github.com/cculianu/Fulcrum/blob/master/README.md#requirements
[4] https://github.com/bitcoin/bitcoin/pull/20702